Begin combining common steps first app test.
