+++ tests/functional/configuration/config.py
@@ -65,6 +65,7 @@ OAUTH_PROXY = ENV['oauth_proxy']
 SPINE_HOSTNAME = (
     # This value is the url returned in the patients response payload which reflects a spine environment.
     # internal-qa environment points to spine int environment.
+    # TODO: Move to env variables.
     "https://veit07.api.service.nhs.uk" if ENVIRONMENT == "internal-dev" else "https://int.api.service.nhs.uk"
 )
